\name{chrom.lookup-package}
\alias{chrom.lookup-package}
\alias{chrom.lookup}
\docType{package}
\title{
\packageTitle{chrom.lookup}
}
\description{
\packageDescription{chrom.lookup}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{chrom.lookup}
\packageIndices{chrom.lookup}
%% ~~ An overview of how to use the package, including the most important ~~
%% ~~ functions ~~
}
\author{
\packageAuthor{chrom.lookup}

Maintainer: \packageMaintainer{chrom.lookup}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\keyword{ chromosome }
\keyword{ genes }
\seealso{
	\code{ \link{ gene.chrom } }
	\code{ \link{ gene.chrom.dict } }
	\code{ \link{ index.chromosome } }
}
