\name{gene.chrom}
\alias{gene.chrom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
The cromosome where the gene belongs according to biomart
}
\description{
Given a gene name, returns the name of the chromosome where it belongs.
}
\usage{
gene.chrom(gene_name, gene_dict)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gene_name}{
A 'string' containing the name of the given gene or 'NA'.
}
  \item{gene_dict}{
A dictionary generated by the function \code{gene.chrom.dict}
}
}
\details{
}
\value{
The name of the chromosome (sting).
}
\references{
}
\author{
}
\note{
}

\seealso{
\code{ \link{ chrom.lookup } }
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (gene_name, gene_dict) 
{
    if (has.key(gene_name, gene_dict)) {
        return(gene_dict[[gene_name]])
    }
    else {
        return(NA)
    }
  }
}

\keyword{ gene }
\keyword{ chromosome }
