\name{chrom.lookup-package}
\alias{chrom.lookup-package}
\alias{chrom.lookup}
\docType{package}
\title{
\packageTitle{chrom.lookup}
}
\description{
\packageDescription{chrom.lookup}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{chrom.lookup}
\packageIndices{chrom.lookup}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{chrom.lookup}

Maintainer: \packageMaintainer{chrom.lookup}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
